# fix: process parallel events