## Motivation

#94 - Mutli-threaded tracing drops most of the events

## Solution

I have reorganized the code so that the lock on extensions is held, but not while `record` is called which should solve the deadlock originally solved in #59 

### Benchmark

I've used the benchmark from #93, please ignore the "filtered" vs "non_filtered" distinction, for the purpose of this PR they should be identical cases.

#### Current `HEAD`

```
otel_many_events/spec_baseline
                        time:   [20.257 µs 20.309 µs 20.370 µs]
                        change: [+2.9172% +3.2204% +3.5153%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 16 outliers among 100 measurements (16.00%)
  7 (7.00%) high mild
  9 (9.00%) high severe
otel_many_events/no_data_baseline
                        time:   [48.353 µs 48.388 µs 48.430 µs]
                        change: [-2.5926% -2.2458% -1.9098%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 13 outliers among 100 measurements (13.00%)
  5 (5.00%) high mild
  8 (8.00%) high severe
otel_many_events/data_only_baseline
                        time:   [67.881 µs 67.920 µs 67.967 µs]
                        change: [-1.5043% -1.3231% -1.1636%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  5 (5.00%) high mild
  7 (7.00%) high severe
otel_many_events/full_filtered
                        time:   [271.34 µs 271.43 µs 271.52 µs]
                        change: [+196.93% +197.43% +197.89%] (p = 0.00 < 0.05)
                        Performance has regressed.
Found 11 outliers among 100 measurements (11.00%)
  3 (3.00%) low severe
  3 (3.00%) low mild
  3 (3.00%) high mild
  2 (2.00%) high severe
otel_many_events/full_not_filtered
                        time:   [256.62 µs 256.89 µs 257.21 µs]
                        change: [-6.6735% -6.2952% -5.9302%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 17 outliers among 100 measurements (17.00%)
  1 (1.00%) high mild
  16 (16.00%) high severe

```

#### This PR

```
otel_many_events/spec_baseline
                        time:   [19.975 µs 20.005 µs 20.040 µs]
                        change: [-1.7956% -1.4097% -1.0523%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  2 (2.00%) high mild
  7 (7.00%) high severe
otel_many_events/no_data_baseline
                        time:   [48.774 µs 48.808 µs 48.847 µs]
                        change: [+0.6351% +0.9073% +1.1618%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 13 outliers among 100 measurements (13.00%)
  3 (3.00%) high mild
  10 (10.00%) high severe
otel_many_events/data_only_baseline
                        time:   [68.291 µs 68.357 µs 68.446 µs]
                        change: [+0.5841% +0.8022% +1.0528%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 12 outliers among 100 measurements (12.00%)
  3 (3.00%) high mild
  9 (9.00%) high severe
otel_many_events/full_filtered
                        time:   [233.29 µs 233.66 µs 234.23 µs]
                        change: [-14.110% -13.925% -13.695%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 14 outliers among 100 measurements (14.00%)
  4 (4.00%) low severe
  3 (3.00%) low mild
  1 (1.00%) high mild
  6 (6.00%) high severe
otel_many_events/full_not_filtered
                        time:   [238.40 µs 238.47 µs 238.55 µs]
                        change: [-7.7942% -7.4832% -7.2150%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 9 outliers among 100 measurements (9.00%)
  3 (3.00%) low severe
  2 (2.00%) low mild
  4 (4.00%) high severe
  ```