@mladedav 2024-02-26
@jtescher do you think you'll have some time to take a look at this? I'd like to try and tackle the span limits next but I think it should be built upon this so I'd like to have even a soft go-ahead on this first.